Move commands back to commands dir to fix auto-discovery issue
